pyproject.toml
[project]
name = "nonebot-plugin-mcversion"
version = "0.2.2"
description = "NoneBot2 plugin for CheckMCupdate"
authors = [
    {name = "CN171-1", email = "3428166361@qq.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.0rc2", "nonebot-adapter-onebot>=2.1.5", "nonebot-plugin-apscheduler>=0.2.0"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/CN171-1/nonebot-plugin-mcversion"
Repository = "https://github.com/CN171-1/nonebot-plugin-mcversion"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"